split-ser: All working still
